Enable subpixel AA for tabs, status-bar etc.
Problem: The GPU layer added to the .gutter .tiles cause the surounding UI to also composite layers, like the tas or the status-bar. This disables subpixel anti-aliasing and text doesn't look that sharp.

Eventough the `.gutter` has `overflow: hidden`, it seems Chrome still thinks there is a chance of overlapping and thus creates the extra layers.

Solution: Creating an own stacking context for the `.gutter` seems to fix it.

Issue #7904