Add `activateItem` config to `Workspace::open`
Now that we have deleted `Project::open(...)`, all the methods that previously
used it need to switch to `Workspace::open(...)` instead. These two methods,
however, are not idempotent because the latter performs additional operations.

Since some specs relied on not activating an item after opening it, this commit
adds a parameter to prevent such activation.

Refs: https://github.com/atom/tabs/commit/1ebcf166c73a9d55932a78b1246dbd5a50077ec5