:bug: Fix folds not preserved when moving multiple selections down
When two or more selections span the rows immediately before a fold,
the `did-change` event dispatched on the deletion will trigger a merge
of the selections, which in turn trigger an unfold at the buffer
position of the new selection, which is now the position of the fold.
Consolidating the selections at the begin of the transaction will
prevent the merge and will keep the fold untouched.