Fix F11 keymaps
Function key keymaps don't work when they feature upperface 'F' characters.
They still tend to work because they typically have an application menu item
which is associated with it and falled back upon. The main difference is that
the originalEvent (KeyboardEvent) is not retained when doing this fallback as
Atom thinks that the menu item was pressed instead of the keymap.

Fixes #10287