Delete specs testing previous behavior
When emitting the `did-change` event, `updateFoldableStatus` used to extend the
change region up and down to include all the lines that changed their
foldability status because of a buffer change. I assume this was supposed to
invalidate folds that *contained* the change whenever a line was edited in a way
that affected also the previous or subsequent ones.

That information, however, is not being used by `DisplayBuffer`, which does not
alter existing folded regions when they become invalid.

I believe the correct behavior should be to unfold the invalid region and
recompute those screen lines. Nonetheless, it seems reasonable to me to keep the
original (wrong) behavior and not address it in this branch, because it strays
from the original intent of this refactoring.

We should probably fix it once for all in another PR or when
integrating/implementing `DisplayLayer`.