Always strip git+ prefix and .git suffix from package repository URLs
Previously, a guard based on the presence of the _id field (which is
inserted by npm during installation) prevented a regex replacement of
the git+ prefix on URLs. Now we always do this. Since the .git suffix
also causes problems and we’re removing that in packages, I now remove
that as well.